Better file split
