+++ src/blockchain/inherent/index.ts
@@ -5,6 +5,8 @@ import { HexString } from '@polkadot/util/types'
 import { getCurrentTimestamp, getSlotDuration } from '../../utils/time-travel'
 
 export { SetValidationData } from './parachain/validation-data'
+export { SetBabeRandomness } from './parachain/babe-randomness'
+export { SetNimbusAuthorInherent } from './parachain/nimbus-author-inherent'
 
 export interface CreateInherents {
   createInherents(parent: Block, params?: BuildBlockParams['inherent']): Promise<HexString[]>
